Authenticator added

+++ controller/book.controller.js
@@ -4,6 +4,7 @@ const bookModel = require("../models/book.model");
 const createBook = async (req, res) => {
 	try {
 		const { title, image, author, genre, description } = req.body;
+		const userId = req.user?._id;
 
 		if (!title || !image || !author || !genre) {
 			throw new Error("Title, image, author and genre are required");
@@ -15,6 +16,7 @@ const createBook = async (req, res) => {
 			author,
 			genre,
 			description,
+			user: userId,
 		});
 
 		res.status(201).json(book);
@@ -25,7 +27,7 @@ const createBook = async (req, res) => {
 
 const getAllBooks = async (req, res) => {
 	try {
-		const books = await bookModel.find({});
+		const books = await bookModel.find({}).populate(user).exec();
 		res.status(200).json(books);
 	} catch (error) {
 		res.status(400).json({ error: error.message });
@@ -39,7 +41,7 @@ const getABook = async (req, res) => {
 			throw new Error("Invalid book id");
 		}
 
-		const book = await bookModel.findById(bookId);
+		const book = await bookModel.findById(bookId).populate(user).exec();
 
 		if (!book) {
 			throw new Error("Book not found");
@@ -54,6 +56,7 @@ const updateABook = async (req, res) => {
 	try {
 		const { bookId } = req.params;
 		const { title, image, author, genre, description } = req.body;
+		const userId = req.user?._id;
 
 		if (!mongoose.Types.ObjectId.isValid(bookId)) {
 			throw new Error("Invalid book id");
@@ -62,6 +65,9 @@ const updateABook = async (req, res) => {
 		const book = await bookModel.findById(bookId);
 		if (!book) {
 			throw new Error("Book not found");
+		}
+		if (book.user.toString() !== userId.toString()) {
+			throw new Error("Forbidden");
 		}
 
 		const updatedBook = await bookModel.findByIdAndUpdate(
@@ -85,11 +91,24 @@ const updateABook = async (req, res) => {
 const deleteABook = async (req, res) => {
 	try {
 		const { bookId } = req.params;
+		const userId = req.user?._id;
+
 		if (!mongoose.Types.ObjectId.isValid(bookId)) {
 			throw new Error("Invalid book id");
 		}
 
-		const book = await bookModel.findByIdAndDelete(bookId);
+		const book = await bookModel.findById(bookId);
+
+		if (!book) {
+			throw new Error("Book not found");
+		}
+
+		if (book.user.toString() !== userId.toString()) {
+			throw new Error("Forbidden");
+		}
+
+		await bookModel.findByIdAndDelete(bookId);
+
 		res.status(200).json({ message: "Book deleted successfully", book });
 	} catch (error) {
 		res.status(400).json({ error: error.message });
